Implemented Kalos & Alolan Pokemon, Pokemon TM Skills, Hoenn Assets & Simulator Overhaul
- Kalos and Alolan Pokemon implemented
- TM lists implemented for Kanto, Alolan Kanto, Johto, Kalos and Alolan Pokemon
- SImulator is now one page. It displays routes and has the option to go to gyms
- Elite 4 for Kanto is implemented after receiving 8 Kanto Badges
- Johto gym leaders are implemented
- Implemented Revelation Dance & First Impression skills
- Added more trainer assets
- Added Hoenn Pokemon assets

+++ PokemonToronto/core/src/com/pokemon/toronto/game/com/pokemon/toronto/Pokemon/RoutePokemon.java
@@ -8,8 +8,8 @@ public class RoutePokemon {
     private int pokemonId;
     private int lowestLevel;
     private int highestLevel;
-    private int rate;
-    public RoutePokemon(int pokemonId, int lowestLevel, int highestLevel, int rate) {
+    private double rate;
+    public RoutePokemon(int pokemonId, int lowestLevel, int highestLevel, double rate) {
         this.pokemonId = pokemonId;
         this.lowestLevel = lowestLevel;
         this.highestLevel = highestLevel;
@@ -28,7 +28,7 @@ public class RoutePokemon {
         return highestLevel;
     }
 
-    public int getRate() {
+    public double getRate() {
         return rate;
     }
